requirements.txt
fastapi==0.95.2
uvicorn[standard]~=0.32.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.6   # Use if you’re connecting to PostgreSQL
mysqlclient==2.1.1       # Use if you’re connecting to MySQL
python-jose==3.3.0       # For JWT token handling
passlib[bcrypt]==1.7.4   # For password hashing
pydantic==1.10.2         # For data validation

dnspython~=2.7.0
idna~=3.10
cryptography~=43.0.3
anyio~=4.6.2.post1
h11~=0.14.0
pip~=23.2.1
rsa~=4.9
pyasn1~=0.6.1
ecdsa~=0.19.0
sniffio~=1.3.1
typing_extensions~=4.12.2
six~=1.16.0
starlette~=0.27.0
email_validator~=2.2.0
pydantic_core~=2.23.4
websockets~=13.1
click~=8.1.7
httptools~=0.6.4
watchfiles~=0.24.0
greenlet~=3.1.1
annotated-types~=0.7.0
python-multipart~=0.0.12
PyYAML~=6.0.2
PyMySQL~=1.1.1
python-dotenv~=1.0.1